3단계 - 양방향을 단방향 (#44)
* refactor: domain layer refactoring

- NoArgumentConstructor public -> protected 변경
- 메뉴 상품 일급컬렌션 수정
- 테이블 단제 지정 생성자를 통해 지정

* refactor: wrap quantity

- 수량 원시값 객체 포장
- jpa converter 추가
- converter auto apply true 추가
- 패키지 분리 infra -> generic

* refactor: menu product 분리

- 메뉴에 포함된 메뉴 상품 도메인 분리
- 메뉴 상품과 상품 서비스 레이어 관계 추가
- dto 수정

* refactor: table, tablegroup 분리

- 테이블, 단체지정 단방향 분리
- 단체 지정 해제 조건 Validator 컴포넌트 추가

* refactor: order, orderLineItem 도메인 분리

- order에 속한 item 양방향 관계 제거
- 서비스 레이어 분리
- response dto 추가

* refatcor: menu domain repackage

- 메뉴, 메뉴그룹, 메뉴상품 리패키징

* refactor: code cleaning

- 메소드, 변수 이름 변경
- stream function 구분

+++ src/test/java/kitchenpos/menu/util/MenuRequestBuilder.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public class MenuRequestBuilder {
     private String name;
-    private int price;
+    private long price;
     private Long menuGroupId;
     private final List<MenuProductRequest> menuProducts = new ArrayList<>();
 
@@ -17,7 +17,7 @@ public class MenuRequestBuilder {
         return this;
     }
 
-    public MenuRequestBuilder withPrice(int price) {
+    public MenuRequestBuilder withPrice(long price) {
         this.price = price;
         return this;
     }
@@ -27,7 +27,7 @@ public class MenuRequestBuilder {
         return this;
     }
 
-    public MenuRequestBuilder addMenuProduct(long productId, int quantity) {
+    public MenuRequestBuilder addMenuProduct(long productId, long quantity) {
         menuProducts.add(new MenuProductRequest(productId, quantity));
         return this;
     }
